Initial commit
Very basic, probably does not yet work. Honestly gradle is giving me a
headache when developing with it, so I might try out maven instead
(given that we are already using a low amount of gradle automatisation).
Only issue is that gradle does simplify the maven publishing process for
me.