Replace Fabric AW impl with our own.
This gives us far more flexibillity so that I should be able to
do the next planned step far more easily. Also increases usabillity
of this plugin in case the AW and my maven repo vanish. Not that I expect
anyone to use it now, but who knows what people are going to do in a few years time.

Perhaps the 4.9 galimulator days will be seen as the golden age of modding
 (for a reason that may be elusive to us now) and people are going to use this
to get in the mood. However unless I fail with SML1 I do not think that
this can happen.

On a sidenote, it will be interesting how SML1 will look like; right now I have
a mod loader planned where individual class files are recompiled and linked to the
rest of the remaining (unmodified) class files. I also have a monolithic mod loader
in mind, which is why I rebated between using "Monolithic Starloader" or "SML1"
as the name. Or perhaps I'll go with "Geolykt's patchset", but things will be a lot
different in a few weeks. To some degree it is interesting to think about what the far
future (i. e. 2027+) has in store for us. Perhaps galim modding will never be a thing
due to climate change or some other apocaliptic event, perhaps it will be a sprawling
community of dozen of people actively collaborating. Maybe I die in a car crash by then.

It is also interesting to debate whether someone (i. e. you) will come along and read this commit,
after all git wasn't meant for posting such message to the future. On the other hand,
it might be the perfect place to do it as preserving software might be more interesting than
other things. Though I have no idea who would have interest in preserving ObfTools for
generations to come.